+++ src/interactors/buildComment.js
@@ -7,9 +7,11 @@ module.exports = ({
   table,
   org,
   repos,
+  isSponsor,
   periodLength,
 }) => {
   const sources = buildSources({ buildGithubLink, org, repos });
   const message = t('table.subtitle', { sources, count: periodLength });
-  return `## ${t('table.title')}\n${message}:\n${table}`;
+  const footer = isSponsor ? '' : `\n${t('table.footer')}`;
+  return `## ${t('table.title')}\n${message}:\n${table}${footer}`;
 };